NEW: Added the ability for each sourcecontrol to control it's representation all the way out to the presentation layer. The idea of this is acctually stolen from isMobile best practice, but not secret :-) The defenition of the method getModifications() has now change. The contract now say that you either return a List of Modification's or a List of org.jdom.Element's. ModificationSet then checks to see wichever it is. It also works in mixed mode, using different sourcecontrols within the same modificationset. NOTE: Sort is removed from modificationset now, but I will fix this in the stylesheet instead (where it probably should be anyway, CC istelf doesn't care). Etc, etc.
TODO: Still not commited, but allready working is also a complementary presentation solution. Sort of a packaging feature that comes with a sourcecontrol (if you like). There are defaults and this change is backwards competable for all elements except Perforce (for now). But this only means that perforce needs it own XLS setup (on parts only). This is expected to be customized by each user anyway so it's not really a problem. Expect this to be commited in a couple of days or so.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@513 878de816-9c12-0410-b5fe-b34e1928e20d