Allow preconfiguration of nested elements! Tested it with root & projects tlugins and global & project property resolution.
The implementation was eased by the following properties of the current implementation:
- validate() is only called when the plugin is given to the user. If it had been called earlier, that would have created some issues.

For the curious the implementation manipulates the DOM. Hence when I find a plugin, I register it (map name & class), and modifiy a copy of the DOM so that

 <plugin name="plugname" class="xxx".,..>
   <....>
 </plugin>

is changed into

 <plugname .,..>
   <....>
 </plugname>

Then I can reuse all existing configuration code.

We also remove the defaultProperties as they become redundant.

Note: it is not possible to preconfigure a plugin argument of name 'name' or 'class' for obvious reasons. We should update the doc for that.

Patch by Jerome Lacoste (jerome@coffeebreaks.org)


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@2061 878de816-9c12-0410-b5fe-b34e1928e20d