Provide extra instructions in failure message
testGetPluginCSS depends on gendoc.css existing in the classpath,
but currently this file is only created by the copy-css Ant target.
If this test is run in a clean checkout before any command line builds
(e.g. when a developer checks out an initial copy of the source tree
and tries to run all the tests in their IDE, which is how I discovered
this issue) then it will always fail.  The fix is simply to run a
command line build before trying to run this test another way.

Most users will never see this message, but those who do are likely
doing what I did, so add a note to the assertion failure message
suggesting they try running a command line build before running this
test again.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4629 878de816-9c12-0410-b5fe-b34e1928e20d