Tell users to create env var when test fails
testBuild_DelEnvVar depends on an environment variable named TESTENV being set
but it doesn't matter what this value is set to.  When this test fails because
that environment variable is not set the failure is obvious when looking at the
code (assertNotNull(env.getVariable(TEST_ENVVAR))) but it is not immediately
obvious what this variable should be set to.  In .ant-scripts/build-test.xml
it appears the actual value is not important, so it can be set to anything.

In order to prevent others from having to hunt this information down in the
future we can provide a more descriptive error message informing users that
it is perfectly safe to simply set it to a dummy value.  Most users will never
see this message because the Ant build scripts automatically set this up before
running the tests, so specifically call it out for developers trying to run
this test in their IDE.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4630 878de816-9c12-0410-b5fe-b34e1928e20d