Add reminder to revisit gendoc.css issue later
gendoc.css is currently created by the Ant build.  Its contents are simply the
concatenation of cruisecontrol.css and configxml-gendoc.css.  However, the
purpose of this file is not immediately obvious because the gendoc'ed
configxml.html file still links directly to cruisecontrol.css and
configxml-gendoc.css in its header (see: configxml_html.vm).

I suspect the reason for creating gendoc.css is so that its contents can be
exposed via JMX (see CruiseControlControllerJMXAdaptor#getPluginCSS), however
the purpose of exposing the contents via JMX is not immediately obvious to me,
either.  I think this should be revisited in the future, and if there's still
a need to expose the contents via JMX then I would like to discuss alternative
solutions such as manually merging the two files together, checking that file
into version control, and updating the configxml.html template to directly
refer to that file.  I don't think it's wise to depend on the presence of a
file like gendoc.css which is nested in our source tree (it is copied to
main/src/net/sourceforge/cruisecontrol/gendoc/html/) but not actually checked
into version control, and I don't think it's wise for any tests or production
functionality to depend on this file being present on the classpath.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4631 878de816-9c12-0410-b5fe-b34e1928e20d