Locate and set local Subversion working copy
The svn info tests fail on systems with newer Subversion clients.  It appears
svn info no longer automatically searches parent directories for a valid local
working copy, so when run out of the directory where our Ant build runs the
tests there were build failures.

The simplest work-around is to always walk the directory hierarchy ourselves
and explicitly set the local working copy path when we find the first .svn
entry.  This initial implementation isn't terribly robust, for example it
doesn't even check that the .svn entry is a directory let alone a valid
Subversion directory, but it is good enough as a first step.

This issue also exposes the fact that our svn tests are somehow dependent,
or at least affected by, the fact that CruiseControl itself uses Subversion.
This diff doesn't address that problem, either, although that is something I
believe we should investigate further and improve.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4634 878de816-9c12-0410-b5fe-b34e1928e20d