Add latest version of XStream
We currently have XStream v1.2.2 in our tree, but that version fails to
correctly recognize JDK 7.  This results in several test failures due to
missing default constructors when building with JDK 7 (see:
http://jira.codehaus.org/browse/XSTR-651).

This commit merely adds the latest XStream release to the tree.  It is
not yet hooked up to the builds; that will happen next.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4636 878de816-9c12-0410-b5fe-b34e1928e20d