Add missing dependency on XPP
XStream requires XPP, but I missed this dependency when I first broke out
build-paths.xml for the reporting dashboard.  That omission didn't break
anything because a version of xpp3 was included in reporting/dashboard/lib
and automatically copied into the .war file.  I only noticed when I tried
to remove that older version in preparation for upgrading to the latest
release.


git-svn-id: https://svn.code.sf.net/p/cruisecontrol/code/trunk@4637 878de816-9c12-0410-b5fe-b34e1928e20d